confirmation dialog added

+++ src/app/accounts/account-form/account-form.component.ts
@@ -7,6 +7,8 @@ import { MatInputModule } from '@angular/material/input';
 import { ActivatedRoute, Params, Router, RouterLink } from '@angular/router';
 import { AccountService } from '../../service/account.service';
 import { NgIf } from '@angular/common';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-account-form',
@@ -15,6 +17,7 @@ import { NgIf } from '@angular/common';
     NgIf,
     RouterLink,
     ReactiveFormsModule,
+    ConfirmDialogComponent,
 
     MatFormFieldModule,
     MatInputModule,
@@ -24,39 +27,47 @@ import { NgIf } from '@angular/common';
   templateUrl: './account-form.component.html',
   styleUrl: './account-form.component.scss'
 })
-export class AccountFormComponent implements OnInit{
+export class AccountFormComponent implements OnInit {
   formGroup = new FormGroup({
     'id': new FormControl(0),
     'name': new FormControl(null, Validators.required),
     'balance': new FormControl(null, [Validators.required, Validators.pattern(/^[0-9]+$/)])
   });
-  
+
   editMode = false;
 
   constructor(
-    private accountService: AccountService, 
-    private router: Router, 
-    private route: ActivatedRoute
-    ){}
+    private accountService: AccountService,
+    private router: Router,
+    private route: ActivatedRoute,
+    private dialog: MatDialog,
+  ) { }
 
-    ngOnInit(): void {
-      this.route.params.subscribe((params: Params) =>{
-        if(params['id']){
-          this.editMode = true;
-          this.accountService.getAccountById(+params['id'])
+  ngOnInit(): void {
+    this.route.params.subscribe((params: Params) => {
+      if (params['id']) {
+        this.editMode = true;
+        this.accountService.getAccountById(+params['id'])
           .then(data => this.formGroup.patchValue(data));
-          this.formGroup.controls.balance.disable();
-        }
-      })
-    }
-
-  onSubmit() {
-    this.accountService.saveAccount({...this.formGroup.getRawValue()}).then(() =>
-    this.router.navigate(['../'], { relativeTo:this.route}));
+        this.formGroup.controls.balance.disable();
+      }
+    })
   }
 
-  onDelete(){
-    this.accountService.deleteAccount(this.formGroup.value.id)
-    .then(() => this.router.navigate(['../'], { relativeTo:this.route}));
+  onSubmit() {
+    this.accountService.saveAccount({ ...this.formGroup.getRawValue() }).then(() =>
+      this.router.navigate(['../'], { relativeTo: this.route }));
+  }
+
+  onDelete() {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: { title: "Are you Sure?", body: `Are you sure you want to delete Account <b> ${this.formGroup.value.name} </b>?` }
+    });
+    dialogRef.afterClosed().subscribe(data => {
+      if (data) {
+        this.accountService.deleteAccount(this.formGroup.value.id)
+          .then(() => this.router.navigate(['../'], { relativeTo: this.route }));
+      }
+    });
   }
 }
