+++ src/app/category/category-form/category-form.component.ts
@@ -8,6 +8,8 @@ import { MatSelectModule } from '@angular/material/select';
 import { ActivatedRoute, Params, Router, RouterLink } from '@angular/router';
 import { CategoryService } from '../../service/category.service';
 import { NgIf } from '@angular/common';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-category-form',
@@ -16,6 +18,7 @@ import { NgIf } from '@angular/common';
     NgIf,
     RouterLink,
     ReactiveFormsModule,
+    ConfirmDialogComponent,
 
     MatIconModule,
     MatFormFieldModule,
@@ -26,8 +29,8 @@ import { NgIf } from '@angular/common';
   templateUrl: './category-form.component.html',
   styleUrl: './category-form.component.scss'
 })
-export class CategoryFormComponent implements OnInit{
-  
+export class CategoryFormComponent implements OnInit {
+
   formGroup = new FormGroup<CategoryForm>({
     id: new FormControl(0),
     name: new FormControl(null, Validators.required),
@@ -37,35 +40,43 @@ export class CategoryFormComponent implements OnInit{
 
   constructor(
     private route: ActivatedRoute,
-    private router : Router,
-    private categoryService: CategoryService
-    ){}
+    private router: Router,
+    private categoryService: CategoryService,
+    private dialog: MatDialog,
+  ) { }
 
   ngOnInit(): void {
-    this.route.params.subscribe((params: Params)=>{
-      if(params['id']){
+    this.route.params.subscribe((params: Params) => {
+      if (params['id']) {
         this.editMode = true;
         this.categoryService.getCategoryById(+params['id'])
-        .then(category => this.formGroup.patchValue(category));
+          .then(category => this.formGroup.patchValue(category));
       }
     });
   }
 
-  onSubmit(){
+  onSubmit() {
     this.categoryService.saveCategory(this.formGroup.getRawValue())
-    .then(() => this.router.navigate(['../'],{relativeTo: this.route}));
+      .then(() => this.router.navigate(['../'], { relativeTo: this.route }));
   }
 
-  onDelete(){
+  onDelete() {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: { title: "Are you Sure?", body: `Are you sure you want to delete category <b> ${this.formGroup.value.name} </b>?` }
+    });
+    dialogRef.afterClosed().subscribe(data => {
+      if (data) {
+        this.categoryService.deleteCategory(this.formGroup.value.id)
+          .then(() => this.router.navigate(['../'], { relativeTo: this.route }));
+      }
+    })
     console.log("On Delete", this.formGroup.value.id);
-    this.categoryService.deleteCategory(this.formGroup.value.id)
-    .then(() => this.router.navigate(['../'],{relativeTo: this.route}));
   }
 }
 
-interface CategoryForm{
-  id:FormControl<number | null>,
-  name:FormControl<string | null>,
+interface CategoryForm {
+  id: FormControl<number | null>,
+  name: FormControl<string | null>,
   transactionType: FormControl<string | null>
 
 }
